refactor(contained): break the last cycle in the runtime's own modules
Quality 4795 -> 5076, past main's 5046. Cycles 3 -> 2.

Quality is the geometric mean of five sub-scores, so the smallest one
decides it: acyclicity scored 2500 against redundancy's 9416, and
`10000/(cycles+1)` means each cycle removed is worth more than the last.
That is why this small change is worth 281 points when splitting an
813-line file was worth five.

`k8s <-> k8s_division <-> lifecycle` was one strongly connected component,
so both edges had to go together — breaking either alone measures nothing.

`Runtime`, `LifecycleError` and `_INACTIVE_STATES` move to a new leaf
module, `runtimes.py`. The type describes a runtime of either target, so
it could not honestly live in `lifecycle` (which acts on them) or in `k8s`
(which is one target of two). `lifecycle` re-exports both names, so
existing imports resolve unchanged.

`sidecar_command` and its three constants move from `k8s_division` into
`k8s`, next to `loader_command` — its exact analogue, a container's shell
script living in the module that owns the pod spec. `k8s_division` imports
them back, so no test changed. The split now reads as: `k8s` owns what the
pod manifest contains, `k8s_division` owns the file-drop protocol the
agent speaks.

Two things were measured and declined. Depth (4211) is not this feature's
to fix: the critical chain runs entirely through `cli` and `workflow`, and
no `contained` module appears on it. Modularity fell 4313 -> 4302, because
a new small module lowers average cluster cohesion — the honest cost of
removing the cycle, and a reason not to add further shim modules here.

The two remaining cycles are both in `factory/workflow/` and predate this
branch. They are worth roughly another 400 points to whoever owns them.
